+++ index.html
@@ -11,9 +11,9 @@
     <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700;800&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="stylesheets/styles.min.css">
     <script src="scripts/controls.js" defer></script>
-    <script src="scripts/statistics.js" defer></script>
-    <script src="scripts/portfolio.js" defer></script>
-    <script src="scripts/blog.js" defer></script>
+    <script src="scripts/statistics.js" type="module" defer></script>
+    <script src="scripts/portfolio.js" type="module" defer></script>
+    <script src="./scripts/blog.js" type="module" defer></script>
 </head>
 <body class="main-content">
     <header class="container active" id="home">
