+++ scripts/data/articles.js
@@ -1,0 +1,46 @@
+const articles = [
+    {
+        title: "How to Create Your Own Website",
+        summary: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Doloribus natus voluptas, eos obcaecati recusandae amet?",
+        image: "port6.jpg",
+        url: "#"
+    },
+    {
+        title: "How to Become an Expert in Web Design",
+        summary: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Doloribus natus voluptas, eos obcaecati recusandae amet?",
+        image: "blog1.jpg",
+        url: "#"
+    },
+    {
+        title: "Become a Web Designer in 10 Days",
+        summary: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Doloribus natus voluptas, eos obcaecati recusandae amet?",
+        image: "blog2.jpg",
+        url: "#"
+    },
+    {
+        title: "Debbuging made easy with Web Inspector",
+        summary: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Doloribus natus voluptas, eos obcaecati recusandae amet?",
+        image: "blog3.jpg",
+        url: "#"
+    },
+    {
+        title: "Get started with Web Design and UI Design",
+        summary: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Doloribus natus voluptas, eos obcaecati recusandae amet?",
+        image: "port1.jpg",
+        url: "#"
+    },
+    {
+        title: "This is what you need to know about Web Design",
+        summary: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Doloribus natus voluptas, eos obcaecati recusandae amet?",
+        image: "port3.jpg",
+        url: "#"
+    },
+    /*
+    Objects can be added here. The title of the object is the name of the article.
+    The summary is the text that appears in the paragraph.
+    The image is the name of the image file.
+    The url is a link to the article.
+     */
+];
+
+export { articles };
