+++ scripts/data/skills.js
@@ -1,0 +1,15 @@
+const skills = {
+    "HTML5": "80%",
+    "CSS3": "95%",
+    "JavaScript": "75%",
+    "React": "75%",
+    "NodeJS": "87%",
+    "Python": "70%"
+    /*
+    Skills can be added here.
+    The image use for the skills are from the following website: https://devicon.dev/
+    Please, use the same naming as the website.
+    */
+};
+
+export { skills };
