+++ scripts/portfolio.js
@@ -1,75 +1,6 @@
-(function() {
-    const projects = [
-        {
-            name: "Project #1",
-            image: "port1.jpg",
-            icons: [
-                { url: "#", icon: "fab fa-github" },
-                { url: "#", icon: "fab fa-behance" },
-                { url: "#", icon: "fab fa-youtube" }
-            ]
-        },
-        {
-            name: "Project #2",
-            image: "port2.jpg",
-            icons: [
-                { url: "#", icon: "fab fa-github" },
-                { url: "#", icon: "fab fa-behance" },
-                { url: "#", icon: "fab fa-youtube" }
-            ]
-        },
-        {
-            name: "Project #3",
-            image: "port3.jpg",
-            icons: [
-                { url: "#", icon: "fab fa-github" },
-                { url: "#", icon: "fab fa-behance" },
-                { url: "#", icon: "fab fa-youtube" }
-            ]
-        },
-        {
-            name: "Project #4",
-            image: "port4.jpg",
-            icons: [
-                { url: "#", icon: "fab fa-github" },
-                { url: "#", icon: "fab fa-behance" },
-                { url: "#", icon: "fab fa-youtube" }
-            ]
-        },
-        {
-            name: "Project #5",
-            image: "port5.jpg",
-            icons: [
-                { url: "#", icon: "fab fa-github" },
-                { url: "#", icon: "fab fa-behance" },
-                { url: "#", icon: "fab fa-youtube" }
-            ]
-        },
-        {
-            name: "Project #6",
-            image: "port6.jpg",
-            icons: [
-                { url: "#", icon: "fab fa-github" },
-                { url: "#", icon: "fab fa-behance" },
-                { url: "#", icon: "fab fa-youtube" }
-            ]
-        },
-        {
-            name: "Project #7",
-            image: "port7.jpg",
-            icons: [
-                { url: "#", icon: "fab fa-github" },
-                { url: "#", icon: "fab fa-behance" },
-                { url: "#", icon: "fab fa-youtube" }
-            ]
-        },
-        /*
-        Objects can be added here. The name of the object is the name of the project.
-        The image is the name of the image file.
-        The icons are an array of objects. Each object has a url and an icon from the Font Awesome library.
-         */
-    ]
+import { projects } from "./data/projects.js";
 
+(function() {
     const create = {
         element: function(type, className) {
             const element = document.createElement(type);
