+++ scripts/statistics.js
@@ -1,18 +1,6 @@
-(function() {
-    const skills = {
-        "HTML5": "80%",
-        "CSS3": "95%",
-        "JavaScript": "75%",
-        "React": "75%",
-        "NodeJS": "87%",
-        "Python": "70%"
-        /*
-        Skills can be added here.
-        The image use for the skills are from the following website: https://devicon.dev/
-        Please, use the same naming as the website.
-        */
-    }
+import { skills } from "./data/skills.js";
 
+(function() {
     const create = {
         element: function(type, className) {
             const element = document.createElement(type);
